requirements.txt
amqp==2.6.0
apturl==0.5.2
asgiref==3.2.7
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
Brlapi==0.7.0
celery==4.4.4
certifi==2019.11.28
chardet==3.0.4
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.5
django-appconf==1.0.4
django-autocomplete-light==3.5.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-modal-forms==1.5.0
django-celery==3.3.1
django-celery-beat==2.0.0
django-celery-email==3.0.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-filters==0.2.1
django-jquery==3.1.0
django-timezone-field==4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
idna==2.8
keyring==18.0.1
kombu==4.6.10
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0
python-crontab==2.5.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
sqlparse==0.3.1
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vine==1.3.0
wadllib==1.3.3
xkit==0.0.0
